changed redundant casting

+++ src/main/java/org/cloudifysource/quality/iTests/test/cli/cloudify/cloud/byon/agentrestart/AgentRestartTest.java
@@ -67,7 +67,7 @@ public class AgentRestartTest extends AbstractAgentMaintenanceModeTest {
 		assertTrue("agent machine did not stop as expected.", removed.await(DEFAULT_WAIT_MINUTES, TimeUnit.MINUTES));
 		assertTrue("agent machine was not added as expected.",added.await(DEFAULT_WAIT_MINUTES, TimeUnit.MINUTES));
 		
-		assertTrue("detected agent removed after cluster was stabilized.", !removedAfterStabilized.await((long) 5, TimeUnit.MINUTES));
+		assertTrue("detected agent removed after cluster was stabilized.", !removedAfterStabilized.await(5l, TimeUnit.MINUTES));
 		assertTrue("detected agent added after cluster was stabilized.", addedAfterStabilized.getCount() == 1);
 		
 		assertNumberOfMachines(2);
